+++ app/static/styles/bg-css/purple.css
@@ -1,0 +1,196 @@
+body {
+  background: #4e54c8;
+  background: -webkit-linear-gradient(to left, #8f94fb, #4e54c8);
+  width: 100%;
+}
+
+.bubble {
+  position: absolute;
+  left: var(--bubble-left-offset);
+  bottom: 0%;
+  display: block;
+  width: var(--bubble-radius);
+  height: var(--bubble-radius);
+  border-radius: 50%;
+  animation: rotate-up var(--bubble-float-duration) ease-in infinite;
+}
+.bubble::before {
+  position: absolute;
+  content: "";
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(255, 255, 255, 0.2);
+  animation: rotate-up alternate infinite;
+}
+.bubble:nth-child(0) {
+  --bubble-left-offset: 43vw;
+  --bubble-radius: 5vw;
+  --bubble-float-duration: 8s;
+}
+.bubble:nth-child(1) {
+  --bubble-left-offset: 3vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 10s;
+}
+.bubble:nth-child(2) {
+  --bubble-left-offset: 30vw;
+  --bubble-radius: 3vw;
+  --bubble-float-duration: 12s;
+}
+.bubble:nth-child(3) {
+  --bubble-left-offset: 99vw;
+  --bubble-radius: 7vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(4) {
+  --bubble-left-offset: 35vw;
+  --bubble-radius: 1vw;
+  --bubble-float-duration: 6s;
+}
+.bubble:nth-child(5) {
+  --bubble-left-offset: 25vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 6s;
+}
+.bubble:nth-child(6) {
+  --bubble-left-offset: 2vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(7) {
+  --bubble-left-offset: 29vw;
+  --bubble-radius: 6vw;
+  --bubble-float-duration: 7s;
+}
+.bubble:nth-child(8) {
+  --bubble-left-offset: 54vw;
+  --bubble-radius: 4vw;
+  --bubble-float-duration: 7s;
+}
+.bubble:nth-child(9) {
+  --bubble-left-offset: 36vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(10) {
+  --bubble-left-offset: 86vw;
+  --bubble-radius: 5vw;
+  --bubble-float-duration: 10s;
+}
+.bubble:nth-child(11) {
+  --bubble-left-offset: 67vw;
+  --bubble-radius: 7vw;
+  --bubble-float-duration: 8s;
+}
+.bubble:nth-child(12) {
+  --bubble-left-offset: 59vw;
+  --bubble-radius: 10vw;
+  --bubble-float-duration: 7s;
+}
+.bubble:nth-child(13) {
+  --bubble-left-offset: 21vw;
+  --bubble-radius: 10vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(14) {
+  --bubble-left-offset: 81vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 8s;
+}
+.bubble:nth-child(15) {
+  --bubble-left-offset: 7vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 10s;
+}
+.bubble:nth-child(16) {
+  --bubble-left-offset: 78vw;
+  --bubble-radius: 7vw;
+  --bubble-float-duration: 8s;
+}
+.bubble:nth-child(17) {
+  --bubble-left-offset: 92vw;
+  --bubble-radius: 1vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(18) {
+  --bubble-left-offset: 9vw;
+  --bubble-radius: 9vw;
+  --bubble-float-duration: 12s;
+}
+.bubble:nth-child(19) {
+  --bubble-left-offset: 70vw;
+  --bubble-radius: 4vw;
+  --bubble-float-duration: 12s;
+}
+.bubble:nth-child(20) {
+  --bubble-left-offset: 44vw;
+  --bubble-radius: 5vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(21) {
+  --bubble-left-offset: 72vw;
+  --bubble-radius: 7vw;
+  --bubble-float-duration: 7s;
+}
+.bubble:nth-child(22) {
+  --bubble-left-offset: 37vw;
+  --bubble-radius: 4vw;
+  --bubble-float-duration: 6s;
+}
+.bubble:nth-child(23) {
+  --bubble-left-offset: 30vw;
+  --bubble-radius: 5vw;
+  --bubble-float-duration: 10s;
+}
+.bubble:nth-child(24) {
+  --bubble-left-offset: 4vw;
+  --bubble-radius: 9vw;
+  --bubble-float-duration: 11s;
+}
+.bubble:nth-child(25) {
+  --bubble-left-offset: 86vw;
+  --bubble-radius: 6vw;
+  --bubble-float-duration: 9s;
+}
+.bubble:nth-child(26) {
+  --bubble-left-offset: 34vw;
+  --bubble-radius: 2vw;
+  --bubble-float-duration: 9s;
+}
+.bubble:nth-child(27) {
+  --bubble-left-offset: 28vw;
+  --bubble-radius: 6vw;
+  --bubble-float-duration: 6s;
+}
+.bubble:nth-child(28) {
+  --bubble-left-offset: 51vw;
+  --bubble-radius: 5vw;
+  --bubble-float-duration: 7s;
+}
+.bubble:nth-child(29) {
+  --bubble-left-offset: 76vw;
+  --bubble-radius: 8vw;
+  --bubble-float-duration: 9s;
+}
+.bubble:nth-child(30) {
+  --bubble-left-offset: 80vw;
+  --bubble-radius: 3vw;
+  --bubble-float-duration: 11s;
+}
+
+@keyframes rotate-up {
+  from {
+    transform: translateY(0) rotate(0deg);
+    opacity: 1;
+    border-radius: 0;
+  }
+  to {
+    transform: translateY(-1000px) rotate(720deg);
+    opacity: 0;
+    border-radius: 50%;
+  }
+}
+
+/*# sourceMappingURL=purple.css.map */
