+++ app/templates/index.html
@@ -7,11 +7,127 @@
 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
-
-    <title>Hello, world!</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-+0n0xVW2eSR5OomGNYDnhzAbDsOXxcvSN1TPprVMTNDbiYZCxYbOOl7+AMvyTG2x" crossorigin="anonymous">
+     <link rel="preconnect" href="https://fonts.gstatic.com">
+    <link href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&display=swap"
+        rel="stylesheet">
+    <link rel= "stylesheet" type= "text/css" href= "{{ url_for('static',filename='styles/bg-css/'+random_background) }}">
+    <link lang='sass' rel="stylesheet" href="./static/styles/main.css">
+    <title>Team 3.3.1.2</title>
   </head>
+
   <body>
-    <h1>Hello!</h1>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+      <div class="bubble"></div>
+
+
+      <header>
+          <div class="container-fluid header-pos">
+              <div class="row">
+                  <div class="col text-center">
+                    <h1 class="text-light">Team 2</h1>
+                  </div>
+              </div>
+          </div>
+      </header>
+      <div class="cotainer-fluid mt-2">
+          <div class="row">
+              <div class="col"></div>
+              <div class="col text-center">
+                    <div class="row">
+                        <div class="col">
+                            <img src="http://via.placeholder.com/500" class="rounded-circle img-fluid w-50" alt="...">
+                        </div>
+                    </div>
+                    <div class="row mt-5">
+                        <div class="col"></div>
+                        <div class="col">
+                            <div class="card rounded-3" style="width: 18rem;">
+                              <div class="card-body text-start">
+                                <h5 class="card-title">Name</h5>
+                                <h6 class="card-subtitle mb-3 text-muted">Lastname</h6>
+                                <p class="card-text card-text-align mt-4 mb-4">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, q</p>
+                                <a class="btn btn-portfolio text-light rounded-pill" href="#" role="button">More about me</a>
+                              </div>
+                            </div>
+                        </div>
+                        <div class="col"></div>
+                    </div>
+                </div>
+              <div class="col text-center">
+                    <div class="row">
+                        <div class="col">
+                            <img src="http://via.placeholder.com/500" class="rounded-circle img-fluid w-50" alt="...">
+                        </div>
+                    </div>
+                    <div class="row mt-5">
+                        <div class="col"></div>
+                        <div class="col">
+                            <div class="card rounded-3" style="width: 18rem;">
+                              <div class="card-body text-start">
+                                <h5 class="card-title">Name</h5>
+                                <h6 class="card-subtitle mb-3 text-muted">Lastname</h6>
+                                <p class="card-text card-text-align mt-4 mb-4">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, q</p>
+                                <a class="btn btn-portfolio text-light rounded-pill" href="#" role="button">More about me</a>
+                              </div>
+                            </div>
+                        </div>
+                        <div class="col"></div>
+                    </div>
+                </div>
+              <div class="col text-center">
+                    <div class="row">
+                        <div class="col">
+                            <img src="http://via.placeholder.com/500" class="rounded-circle img-fluid w-50" alt="...">
+                        </div>
+                    </div>
+                    <div class="row mt-5">
+                        <div class="col"></div>
+                        <div class="col">
+                            <div class="card rounded-3" style="width: 18rem;">
+                              <div class="card-body text-start">
+                                <h5 class="card-title">Name</h5>
+                                <h6 class="card-subtitle mb-3 text-muted">Lastname</h6>
+                                <p class="card-text card-text-align mt-4 mb-4">Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, q</p>
+                                <a class="btn btn-portfolio text-light rounded-pill" href="#" role="button">More about me</a>
+                              </div>
+                            </div>
+                        </div>
+                        <div class="col"></div>
+                    </div>
+                </div>
+              <div class="col"></div>
+          </div>
+      </div>
+
 
     <!-- Option 1: Bootstrap Bundle with Popper -->
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js" integrity="sha384-gtEjrD/SeCtmISkJkNUaaKMoLD0//ElJ19smozuHV6z3Iehds+3Ulb9Bn9Plx0x4" crossorigin="anonymous"></script>
